+++ crates/codegen/src/irgen/mod.rs
@@ -7,6 +7,9 @@ pub mod statements;
 /// Argument Call Module
 pub mod arg_calls;
 
+/// Builtin Function Module
+pub mod builtin_function;
+
 /// Prelude wraps common utilities.
 pub mod prelude {
     pub use super::{arg_calls::*, constants::*, statements::*};